refactor: drop Tier 1/Tier 2 terminology; use family names
The Tier 1/Tier 2 classification was never defined in any normative
document — only in source comments in tenferro-ops/src/std_tensor_op.rs
and in the filenames tenferro-ops/src/ad/elementwise_tier{1,2}.rs.
The terms carried implicit hierarchy suggestions ("primary" vs
"secondary") that the code did not actually justify.

Changes:
- Renamed tenferro-ops/src/ad/elementwise_tier1.rs to semiring.rs
  (contents: AD rules for Add/Mul/Neg/Conj — the operations that form
  the semiring-arithmetic family).
- Renamed tenferro-ops/src/ad/elementwise_tier2.rs to elementwise.rs
  (contents: Div/Abs/Sign — elementwise ops outside the semiring).
- Merged linearize_semiring + linearize_non_semiring into a single
  pub fn linearize with one exhaustive match. Same for transpose_rule.
  This removes the artificial two-dispatcher split left over from the
  pre-Stage-2 era when SemiringOp<Alg> existed alongside StdTensorOp.
- Renamed the StdTensorOp enum's category comments in std_tensor_op.rs
  to drop "Tier 1:" / "Tier 2:" prefixes, keeping the family labels
  (Semiring arithmetic core, Elementwise, Analytic, Diagonal, Indexing,
  Reductions).
- Updated the docs/spec/extension-op.md cross-reference to the new
  semiring.rs filename.

Behavior unchanged. All existing AD unit/integration tests pass
(tenferro-ops 59 + 4 doc, tenferro ad 69, symbolic_grad 4,
extension_op 9).
